actualizacion

+++ ciclo3/src/main/java/co/edu/bosque/ciclo3/dao/ProductosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.bosque.ciclo3.model.Productos;
 
-public interface ProductosDAO extends JpaRepository<Productos, Integer> {
+public interface ProductosDAO extends JpaRepository<Productos, Long> {
 
 }
